docs: Comment out UML markdown, use image
Commenting out the url that generates the image based on the UML code. We need to specify the branch name and this can not happen while we are in develop/main. We need to get the branch name first if we continue using two branches. For now, we can use the rendered image instead.