fix: temporarily disable push e2e test on ubuntu
It keeps failing on Ubuntu with the message "cannot reuse body, request must be retried".
This needs a bit more investigation, since the message seems to be coming from inside
`oras.Copy` and this particular test works on Windows.

Disabling the push test on Ubuntu for now to not keep blocking the pipelines.